Hotfix/1.1.0

+++ app/controllers/v1/comments_controller.rb
@@ -1,4 +1,5 @@
 class V1::CommentsController < ApplicationController
+  before_action :authorize, only: [:create, :destroy]
 
   def index
     comments = Comment.where(time_report_id: params[:time_report_id]).order(created_at: :desc)
@@ -8,7 +9,7 @@ class V1::CommentsController < ApplicationController
   end
 
   def create
-    comment = Comment.new(comment_params)
+    comment = current_user.comments.create(comment_params)
     if comment.save
       comment.notice
       render json: comment.to_json(include: { user: { except: [:uid, :email], methods: :avatar_url }},
@@ -18,12 +19,14 @@ class V1::CommentsController < ApplicationController
 
   def destroy
     comment = Comment.includes(:notices).find(params[:id])
-    comment.destroy!
-    render json: comment.to_json(include: { user: { except: [:uid, :email], methods: :avatar_url }},
-      methods: :likes_count)
+    if current_user == comment.user
+      comment.destroy!
+      render json: comment.to_json(include: { user: { except: [:uid, :email], methods: :avatar_url }},
+        methods: :likes_count)
+    end
   end
 
   private def comment_params
-    params.require(:comment).permit(:content, :user_id, :time_report_id)
+    params.require(:comment).permit(:content, :time_report_id)
   end
 end
